+++ beacon-chain/core/helpers/legacy.go
@@ -2,27 +2,20 @@ package helpers
 
 import (
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/state"
-	"github.com/prysmaticlabs/prysm/v5/math"
-	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/runtime/version"
 )
 
-// IsLegacyDepositProcessPeriod determines if the current state should use the legacy deposit process.
-func IsLegacyDepositProcessPeriod(beaconState state.BeaconState, canonicalEth1Data *ethpb.Eth1Data) bool {
-	// Before the Electra upgrade, always use the legacy deposit process.
-	if beaconState.Version() < version.Electra {
-		return true
+// DepositRequestHaveStarted determines if the deposit requests have started
+func DepositRequestHaveStarted(beaconState state.BeaconState) bool {
+	if beaconState.Version() >= version.Electra {
+		requestsStartIndex, err := beaconState.DepositRequestsStartIndex()
+		if err == nil {
+			// deposit_requests_start_index will only be set once
+			// eth1data will be frozen
+			if beaconState.Eth1DepositIndex() == requestsStartIndex {
+				return true
+			}
+		}
 	}
-
-	// Handle the transition period between the legacy and the new deposit process.
-	requestsStartIndex, err := beaconState.DepositRequestsStartIndex()
-	if err != nil {
-		// If we can't get the deposit requests start index,
-		// we should default to the legacy deposit process.
-		return true
-	}
-
-	//canonicalEth1Data should never be nil
-	eth1DepositIndexLimit := math.Min(canonicalEth1Data.DepositCount, requestsStartIndex)
-	return beaconState.Eth1DepositIndex() < eth1DepositIndexLimit
+	return false
 }
