--- beacon-chain/rpc/prysm/v1alpha1/validator/proposer_eth1data_test.go
@@ -1,32 +1,0 @@
-package validator
-
-import (
-	"testing"
-
-	"github.com/prysmaticlabs/prysm/v5/testing/util"
-	"github.com/stretchr/testify/require"
-)
-
-// TestDepositRequestHaveStarted contains several test cases for depositRequestHaveStarted.
-func TestDepositRequestHaveStarted(t *testing.T) {
-	t.Run("Version below Electra returns false", func(t *testing.T) {
-		st, _ := util.DeterministicGenesisStateBellatrix(t, 1)
-		result := depositRequestHaveStarted(st)
-		require.False(t, result)
-	})
-
-	t.Run("Version is Electra or higher, no error, but Eth1DepositIndex != requestsStartIndex returns false", func(t *testing.T) {
-		st, _ := util.DeterministicGenesisStateElectra(t, 1)
-		require.NoError(t, st.SetEth1DepositIndex(1))
-		result := depositRequestHaveStarted(st)
-		require.False(t, result)
-	})
-
-	t.Run("Version is Electra or higher, no error, and Eth1DepositIndex == requestsStartIndex returns true", func(t *testing.T) {
-		st, _ := util.DeterministicGenesisStateElectra(t, 1)
-		require.NoError(t, st.SetEth1DepositIndex(33))
-		require.NoError(t, st.SetDepositRequestsStartIndex(33))
-		result := depositRequestHaveStarted(st)
-		require.True(t, result)
-	})
-}
